requirements.txt
flask==3.0.2
flask-cors==5.0.1
tensorflow==2.19.0
numpy==1.26.4
Pillow==10.2.0
gunicorn==21.2.0
pandas==2.2.1
scikit-learn==1.4.2
opencv-python==4.8.0.76
